Fix construction of ByteSize from bits for non multiple of 8 bits

+++ src/ByteSizeLib/ByteSize.cs
@@ -130,7 +130,7 @@ namespace ByteSizeLib
         {
             Bits = bits;
 
-            Bytes = bits / BitsInByte;
+            Bytes = (double)bits / BitsInByte;
         }
 
         public ByteSize(double bytes)
